Extract requestBuilders into RequestBuilder.kt
